Filter out events, Remove custom output, Add multi file drop.
The first change I made is this does not publicize fields with names that match event names in that type and so it should not have the same issue as the current version.

The second change I made is that it no longer can do custom output locations and names but instead it CAN now drop any number of assemblies on the exe at a time and it will publicize them all and put them all in the publicized_assemblies folder that it creates.